trino_nl2sql/requirements.txt
# Trino NL2SQL Requirements
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
openai==1.109.1
python-dotenv==1.0.1
flask==3.0.0
trino==0.329.0
sqlparse==0.5.0
pandas==2.2.0
tabulate==0.9.0
pyyaml==6.0.1
